+++ src/intentkit/proof_checkers/__init__.py
@@ -1,5 +1,6 @@
 """Extensible, local-first proof-checker support for Intent Kit."""
 
+from .external import ExternalCheckerRegistry, ExternalProcessChecker
 from .models import Artifact, CheckerDescriptor, CheckRequest, CheckResult, CheckState, ProofChecker
 from .registry import CheckerRegistry
 from .runner import ProofRunner
@@ -11,6 +12,8 @@ __all__ = [
     "CheckResult",
     "CheckState",
     "CheckerRegistry",
+    "ExternalCheckerRegistry",
+    "ExternalProcessChecker",
     "ProofChecker",
     "ProofRunner",
 ]